:recycle: api: move jury-session-repository to src
