+++ api/src/certification/session/infrastructure/repositories/jury-session-repository.js
@@ -1,9 +1,9 @@
-import { knex } from '../../bookshelf.js';
-import { fetchPage } from '../../utils/knex-utils.js';
-import { NotFoundError } from '../../../domain/errors.js';
-import { JurySession, statuses } from '../../../domain/models/JurySession.js';
-import { CertificationOfficer } from '../../../../src/certification/session/domain/models/CertificationOfficer.js';
-import { PGSQL_UNIQUE_CONSTRAINT_VIOLATION_ERROR } from '../../../../db/pgsql-errors.js';
+import { knex } from '../../../../../lib/infrastructure/bookshelf.js';
+import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { JurySession, statuses } from '../../../../../lib/domain/models/JurySession.js';
+import { CertificationOfficer } from '../../domain/models/CertificationOfficer.js';
+import { PGSQL_UNIQUE_CONSTRAINT_VIOLATION_ERROR } from '../../../../../db/pgsql-errors.js';
 
 const COLUMNS = Object.freeze([
   'sessions.*',
