+++ api/tests/certification/session/integration/infrastructure/repositories/jury-session-repository_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
-import { databaseBuilder, expect, catchErr, domainBuilder } from '../../../../test-helper.js';
-import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import { JurySession, statuses } from '../../../../../lib/domain/models/JurySession.js';
-import { CertificationOfficer } from '../../../../../src/certification/session/domain/models/CertificationOfficer.js';
-import * as jurySessionRepository from '../../../../../lib/infrastructure/repositories/sessions/jury-session-repository.js';
+import { databaseBuilder, expect, catchErr, domainBuilder } from '../../../../../test-helper.js';
+import { NotFoundError } from '../../../../../../lib/domain/errors.js';
+import { JurySession, statuses } from '../../../../../../lib/domain/models/JurySession.js';
+import { CertificationOfficer } from '../../../../../../src/certification/session/domain/models/CertificationOfficer.js';
+import * as jurySessionRepository from '../../../../../../src/certification/session/infrastructure/repositories/jury-session-repository.js';
 
 describe('Integration | Repository | JurySession', function () {
   describe('#get', function () {
